+++ osmnx/_api.py
@@ -5,17 +5,21 @@
 # Web: https://github.com/gboeing/osmnx
 ################################################################################
 
-from .core import gdf_from_place
-from .core import gdf_from_places
-from .core import graph_from_address
-from .core import graph_from_bbox
-from .core import graph_from_file
-from .core import graph_from_place
-from .core import graph_from_point
-from .core import graph_from_polygon
+from .bearing import add_edge_bearings
+
+from .boundaries import gdf_from_place
+from .boundaries import gdf_from_places
+
+from .distance import get_nearest_edge
+from .distance import get_nearest_edges
+from .distance import get_nearest_node
+from .distance import get_nearest_nodes
 
 from .elevation import add_edge_grades
 from .elevation import add_node_elevations
+
+from .folium import plot_graph_folium
+from .folium import plot_route_folium
 
 from .footprints import footprints_from_address
 from .footprints import footprints_from_place
@@ -23,12 +27,17 @@ from .footprints import footprints_from_point
 from .footprints import footprints_from_polygon
 from .footprints import plot_footprints
 
+from .graph import graph_from_address
+from .graph import graph_from_bbox
+from .graph import graph_from_file
+from .graph import graph_from_place
+from .graph import graph_from_point
+from .graph import graph_from_polygon
+
 from .plot import plot_figure_ground
 from .plot import plot_graph
-from .plot import plot_graph_folium
 from .plot import plot_graph_route
 from .plot import plot_graph_routes
-from .plot import plot_route_folium
 from .plot import plot_shape
 
 from .pois import pois_from_address
@@ -56,12 +65,7 @@ from .utils import config
 from .utils import log
 from .utils import ts
 
-from .utils_geo import add_edge_bearings
 from .utils_geo import geocode
-from .utils_geo import get_nearest_edge
-from .utils_geo import get_nearest_edges
-from .utils_geo import get_nearest_node
-from .utils_geo import get_nearest_nodes
 
 from .utils_graph import gdfs_to_graph
 from .utils_graph import get_undirected
